+++ CHANGELOG.md
@@ -12,4 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - A macro `n` for easily creating real `RefractiveIndexSpec`s.
 - A Petzval lens example.
 
+### Fixed
+
+- The aperture stop calculation of the paraxial view now correctly finds the minimum semi-diameter to pseudo-marginal ray height ratio by first taking the absolute value of the ratios. 
+
 [0.1.0]: https://github.com/kmdouglass/cherry/releases/tag/v0.1.0